Change the icon picture
It's the launch window for Mariner 1 and 2, with the transfer for
Mariner 2 highlighted. Plotted in-game (RSS with Principia) using the
"minimize departure Δv" option - this more-or-less validates the use of
the Porkchop plotter to generate a good trajectory.

The calculation I was most worried about is the "Distance between
bodies", since that was the only calculation not basically copied from
TWP.
Mariner 2 was 57.8 Gm from Earth at the time of the flyby; TWP2 gives a
value of 56.8 Gm. Given that the window was planned in TWP2 two years in
advance, this difference can be safely ascribed to differences between
propagation according to Keplerian osculating elements (TWP2) and full
N-body propagation (reality/Principia) of Venus and the spacecraft.