Ask the user for the central body directly
This is preparation for a later update where the target can be a vessel.
Then, this would act as a filter for valid origin/destination bodies;
this pre-empts the problem where a user would try to select their own
vessel as origin and a planet as target, which would not work for a
reason that may not be obvious to them - namely, that for an
interplanetary trajectory, you'd need to select "Kerbin" or "Earth" as
departure body.