More accurate Δv calculation for a minimal capture
The target orbit is not just a (theoretical) elliptical orbit; it also
needs to have its apoapsis within the SoI of the destination planet.

In some cases, the velocity at apoapsis of such an orbit is larger than
vinf; in that case, the transfer cannot be done and we don't plot it.