Pricipia: assume infinite SoI for Δv calculation
For the stock system, the SoI approach is exact. With Principia, the
gravitational effect of the planet continues to slow the spacecraft even
after it leaves the "SoI".

Previously (and with stock physics), the escape Δv was calculated
assuming the transfer initial velocity was the velocity when reaching
the SoI radius. With this change, with Principia gravity, the transfer
initial velocity is assumed to be the velocity at infinity for the
hyperbolic escape orbit. For an interplanetary transfer starting from
Earth, the difference in Δv is between 30 and 40 m/s.

I did a test mission to Mars before and after this change. Before, the Δv
predicted by TWP was too small by 39 m/s; after, it was exactly right

It looks like this fixes the inaccuracy when using Principia.