Remove trailing space from date field
RSSTimeFormatter's `PrintDateCompact` implementation includes a trailing
space, which is then included in the initial value for the date fields.
This causes parsing to fail if the user makes a change to those fields,
but doesn't remove the space.

With this change, the initial value in the date fields is correct and
parseable.