Fix #1
When a vessel target was set (for example from the Tracking Station
scene), and then a scene change occurs, an InvalidOperationException is
thrown during the OnGUI call (i.e. multiple times per frame). The cause
is that all vessels are destroyed and recreated (if not in the editor)
between scenes; therefore, the EndPoint.Vessel field will be null (due
to Unity's lifecycle check).

To fix this, when the window is opened and either Endpoint is in an
invalid state, reset the inputs.